+++ packages/flutter_driver/lib/src/common/enum_util.dart
