+++ packages/flutter_driver/lib/src/common/error.dart
