+++ packages/flutter_driver/lib/src/common/gesture.dart
