+++ packages/flutter_driver/test/src/extension_test.dart
@@ -3,9 +3,9 @@
 // found in the LICENSE file.
 
 import 'package:flutter/scheduler.dart';
-import 'package:flutter_driver/src/extension.dart';
-import 'package:flutter_driver/src/find.dart';
-import 'package:flutter_driver/src/request_data.dart';
+import 'package:flutter_driver/src/common/find.dart';
+import 'package:flutter_driver/src/common/request_data.dart';
+import 'package:flutter_driver/src/extension/extension.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 void main() {
